refactor: AddEditTranfer form
