+++ frontend/src/views/Transfers/AddEditTransfer/AddEditTransfer.jsx
@@ -1,31 +1,43 @@
-import React, { useState, useEffect } from 'react'
-import { useForm, FormProvider } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
 import { useMutation } from '@tanstack/react-query'
+import { useEffect } from 'react'
+import { FormProvider, useForm } from 'react-hook-form'
 import { useNavigate, useParams } from 'react-router-dom'
 
 // Components
+import BCAlert from '@/components/BCAlert'
+import BCBox from '@/components/BCBox'
 import BCTypography from '@/components/BCTypography'
+import Loading from '@/components/Loading'
 import ProgressBreadcrumb from '@/components/ProgressBreadcrumb'
 import { AddEditTransferSchema } from './_schema'
-import TransferForm from './TransferForm'
-import BCBox from '@/components/BCBox'
-import BCAlert from '@/components/BCAlert'
-import Loading from '@/components/Loading'
 import TransferGraphic from './components/TransferGraphic'
 
 // Hooks
-import { useCurrentUser } from '@/hooks/useCurrentUser'
-import { useApiService } from '@/services/useApiService'
 import { TRANSACTIONS } from '@/constants/routes/routes'
+import { useCurrentUser } from '@/hooks/useCurrentUser'
+import { useRegExtOrgs } from '@/hooks/useOrganization'
+import { useTransfer } from '@/hooks/useTransfer'
+import { useApiService } from '@/services/useApiService'
 import { convertObjectKeys, formatDateToISO } from '@/utils/formatters'
+
+import { Box, Button } from '@mui/material'
+import AgreementDate from './components/AgreementDate'
+import Comments from './components/Comments'
+import SigningAuthority from './components/SigningAuthority'
+import TransferDetails from './components/TransferDetails'
 
 export const AddEditTransfer = () => {
   const navigate = useNavigate()
   const apiService = useApiService()
   const { transferId } = useParams() // This extracts the transferId from the URL
   const { data: currentUser } = useCurrentUser()
-  const [organizations, setOrganizations] = useState([])
+  const { data: orgData } = useRegExtOrgs()
+  const { data: transferData, isFetched } = useTransfer(transferId, {
+    enabled: !!transferId,
+    retry: false
+  })
+  // const [organizations, setOrganizations] = useState([])
 
   const methods = useForm({
     resolver: yupResolver(AddEditTransferSchema),
@@ -47,79 +59,69 @@ export const AddEditTransfer = () => {
    * is loaded. It fetches the transfer data using the provided `transferId`, and then resets the form
    * fields with the fetched data, formatting and handling null values appropriately.
    * In case of an error during the fetch operation, it logs the error to the console.
-  */
+   */
   useEffect(() => {
-    const fetchTransferData = async () => {
-      if (!transferId) return
-  
-      try {
-        const response = await apiService.get(`/transfers/${transferId}`)
-        const transferData = response.data
-  
-        // Populate the form with fetched transfer data
-        methods.reset({
-          fromOrganizationId: transferData.from_organization.organization_id,
-          toOrganizationId: transferData.to_organization.organization_id,
-          quantity: transferData.quantity,
-          pricePerUnit: transferData.price_per_unit,
-          signingAuthorityDeclaration: transferData.signing_authority_declaration,
-          comments: transferData.comments.comment, // Assuming you only want the comment text
-          agreementDate: transferData.agreement_date ? 
-            new Date(transferData.agreement_date).toISOString().split('T')[0] : 
-            new Date().toISOString().split('T')[0], // Format date or use current date as fallback
-        })
-      } catch (error) {
-        console.error('Error fetching transfer data:', error)
-      }
+    if (!transferId) return
+    if (isFetched && transferData) {
+      // Populate the form with fetched transfer data
+      methods.reset({
+        fromOrganizationId: transferData.from_organization.organization_id,
+        toOrganizationId: transferData.to_organization.organization_id,
+        quantity: transferData.quantity,
+        pricePerUnit: transferData.price_per_unit,
+        signingAuthorityDeclaration: transferData.signing_authority_declaration,
+        comments: transferData.comments.comment, // Assuming you only want the comment text
+        agreementDate: transferData.agreement_date
+          ? new Date(transferData.agreement_date).toISOString().split('T')[0]
+          : new Date().toISOString().split('T')[0] // Format date or use current date as fallback
+      })
     }
-    if (transferId) fetchTransferData()
-  }, [transferId])
-  
+  }, [isFetched, transferId])
+
   /**
    * Fetch the list of registered external organizations
-  */ 
-  useEffect(() => {
-    const fetchOrganizations = async () => {
-      try {
-        const response = await apiService.get(
-          '/organizations/registered/external'
-        )
-        const orgs = response.data.map((org) => ({
-          value: parseInt(org.organization_id),
-          label: org.name || 'Unknown'
-        }))
-        setOrganizations(orgs)
-      } catch (error) {
-        console.error('Error fetching organizations:', error)
-      }
-    }
-    fetchOrganizations()
-  }, [])
+   */
+  // useEffect(() => {
+  //   const fetchOrganizations = async () => {
+  //     try {
+  //       const response = await apiService.get(
+  //         '/organizations/registered/external'
+  //       )
+  //       const orgs = response.data.map((org) => ({
+  //         value: parseInt(org.organization_id),
+  //         label: org.name || 'Unknown'
+  //       }))
+  //       setOrganizations(orgs)
+  //     } catch (error) {
+  //       console.error('Error fetching organizations:', error)
+  //     }
+  //   }
+  //   fetchOrganizations()
+  // }, [])
 
   const { mutate, isLoading, isError } = useMutation({
-      mutationFn: (convertedPayload) => {
-        if (transferId) {
-          // If editing, use PUT request
-          return apiService.put(`/transfers`, convertedPayload)
-        } else {
-          // If adding new, use POST request
-          return apiService.post('/transfers', convertedPayload)
+    mutationFn: (convertedPayload) => {
+      if (transferId) {
+        // If editing, use PUT request
+        return apiService.put(`/transfers`, convertedPayload)
+      } else {
+        // If adding new, use POST request
+        return apiService.post('/transfers', convertedPayload)
+      }
+    },
+    onSuccess: (response) => {
+      // Redirect on success
+      navigate(TRANSACTIONS, {
+        state: {
+          message: 'Transfer successfully submitted.',
+          severity: 'success'
         }
-      },
-      onSuccess: (response) => {
-        // Redirect on success
-        navigate(TRANSACTIONS, {
-          state: {
-            message: 'Transfer successfully submitted.',
-            severity: 'success',
-          },
-        })
-      },
-      onError: (error) => {
-        console.error('Error submitting transfer:', error)
-      },
+      })
+    },
+    onError: (error) => {
+      console.error('Error submitting transfer:', error)
     }
-  )
+  })
 
   const handleSubmitForm = (form) => {
     form.fromOrganizationId = parseInt(form.fromOrganizationId)
@@ -138,8 +140,9 @@ export const AddEditTransfer = () => {
   const quantity = parseInt(watch('quantity'))
   const creditsFrom = currentUser?.organization?.name
   const creditsTo =
-    organizations.find((org) => org.value === watch('toOrganizationId'))
-      ?.label || ''
+    orgData.find(
+      (org) => parseInt(org.organization_id) === watch('toOrganizationId')
+    )?.label || ''
   const pricePerUnit = watch('pricePerUnit')
   const totalValue =
     quantity && pricePerUnit ? parseInt(quantity * pricePerUnit) : 0
@@ -190,10 +193,25 @@ export const AddEditTransfer = () => {
             data-testid="new-transfer-form"
           >
             {/* Transfer Form Fields */}
-            <TransferForm
-              currentOrg={currentUser?.organization}
-              organizations={organizations}
-            />
+            <TransferDetails />
+
+            <AgreementDate />
+
+            <Comments />
+
+            <SigningAuthority />
+
+            {/* Save Draft Button */}
+            <Box mt={2} display="flex" justifyContent="flex-end">
+              <Button
+                type="submit"
+                variant="contained"
+                color="primary"
+                onClick={() => console.log('Submitting form')}
+              >
+                Save Draft
+              </Button>
+            </Box>
           </form>
         </FormProvider>
       </BCBox>
