+++ frontend/src/views/Transfers/AddEditTransfer/components/Comments.jsx
@@ -1,13 +1,16 @@
-import React, { useState } from 'react'
-import { Box, Collapse, IconButton, TextField, Typography } from '@mui/material'
+import { Box, Collapse, IconButton, TextField } from '@mui/material'
+import { useState } from 'react'
 
 // MUI Icons
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
 import ExpandLessIcon from '@mui/icons-material/ExpandLess'
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
+import { useFormContext } from 'react-hook-form'
 import LabelBox from './LabelBox'
 
-const Comments = ({ register }) => {
+const Comments = () => {
   const [isExpanded, setIsExpanded] = useState(false)
+
+  const { register } = useFormContext()
 
   const handleToggle = () => {
     setIsExpanded(!isExpanded)
