+++ frontend/src/views/Transfers/AddEditTransfer/components/TransferDetails.jsx
@@ -1,29 +1,76 @@
-import React from 'react'
+import BCBox from '@/components/BCBox'
+import { useCurrentUser } from '@/hooks/useCurrentUser'
+import { useRegExtOrgs } from '@/hooks/useOrganization'
 import {
   FormControl,
-  Select,
   MenuItem,
+  Select,
   TextField,
   Typography
 } from '@mui/material'
-import BCBox from '@/components/BCBox'
+import { useEffect, useState } from 'react'
+import { Controller, useFormContext } from 'react-hook-form'
 import LabelBox from './LabelBox'
-import { Controller } from 'react-hook-form'
 
-const TransferDetails = ({
-  register,
-  errors,
-  organizations,
-  currentOrg,
-  totalValue,
-  control,
-  renderError
-}) => {
+const calculateTotalValue = (quantity, pricePerUnit) => {
+  const quantityNum = parseFloat(quantity)
+  const priceNum = parseFloat(pricePerUnit)
+  return !isNaN(quantityNum) && !isNaN(priceNum) ? quantityNum * priceNum : 0
+}
+
+const TransferDetails = () => {
+  const [totalValue, setTotalValue] = useState(0.0)
+  const {
+    register,
+    control,
+    watch,
+    formState: { errors }
+  } = useFormContext()
+  const { data: currentUser } = useCurrentUser()
+  const { data: orgData } = useRegExtOrgs()
+  const organizations = orgData?.map((org) => ({
+    value: parseInt(org.organization_id),
+    label: org.name || 'Unknown'
+  }))
+
+  const quantity = watch('quantity')
+  const pricePerUnit = watch('pricePerUnit')
+
+  useEffect(() => {
+    const newTotalValue = calculateTotalValue(quantity, pricePerUnit)
+    if (totalValue !== newTotalValue) {
+      setTotalValue(newTotalValue)
+    }
+  }, [quantity, pricePerUnit])
+
+  const renderError = (fieldName, sameAsField = null) => {
+    // If the sameAsField is provided and is true, hide errors for this field
+    if (sameAsField && watch(sameAsField)) {
+      return null
+    }
+
+    return (
+      errors[fieldName] && (
+        <Typography
+          color="error"
+          variant="caption"
+          sx={{
+            marginLeft: '14px',
+            marginRight: '14px',
+            marginTop: '4px',
+            marginBottom: '-20px'
+          }}
+        >
+          {errors[fieldName].message}
+        </Typography>
+      )
+    )
+  }
   return (
     <BCBox>
       <LabelBox label="Transfer Details (required)">
         <Typography variant="body1" component="div">
-          {`${currentOrg?.name} transfers `}
+          {`${currentUser?.organization?.name} transfers `}
           <TextField
             {...register('quantity')}
             placeholder="Quantity"
@@ -70,11 +117,17 @@ const TransferDetails = ({
                     }
                   }}
                   renderValue={(selected) => {
-                    if (selected === "") {
-                      return <em>Select an Organization</em>;
+                    if (selected === '') {
+                      return <em>Select an Organization</em>
                     }
-                    const selectedOrg = organizations.find(org => org.value === selected);
-                    return selectedOrg ? selectedOrg.label : <em>Select an Organization</em>;
+                    const selectedOrg = organizations.find(
+                      (org) => org.value === selected
+                    )
+                    return selectedOrg ? (
+                      selectedOrg.label
+                    ) : (
+                      <em>Select an Organization</em>
+                    )
                   }}
                 >
                   <MenuItem value="">
